Refactor sim diretory
