+++ runtime_lib/test_library.h
@@ -57,11 +57,6 @@ static uint16_t nextAlignedAddr;
 #define MODE_PL 1
 #define MODE_MEM 2
 
-//#define HIGH_ADDR(addr)	((addr & 0xffffffff00000000) >> 32)
-//#define LOW_ADDR(addr)	(addr & 0x00000000ffffffff)
-
-//#define MLIR_STACK_OFFSET 4096
-
 struct aie_libxaie_ctx_t {
   XAieGbl_Config *AieConfigPtr;
   XAieGbl AieInst;
